Feature: Async adapter with config flag
