+++ socketio-server/Cargo.toml
@@ -17,3 +17,7 @@ tower-http = "0.4.0"
 tracing = "0.1.37"
 thiserror = "1.0.40"
 itertools = "0.10.5"
+
+
+[features]
+async_adapter = []
