+++ socketio-server/src/operators.rs
@@ -1,11 +1,11 @@
-use std::{pin::Pin, sync::Arc, time::Duration};
+use std::{collections::HashSet, pin::Pin, sync::Arc, time::Duration};
 
 use futures::Stream;
 use itertools::Itertools;
 use serde::{de::DeserializeOwned, Serialize};
 
 use crate::{
-    adapter::{Adapter, BroadcastFlags, BroadcastOptions, Room},
+    adapters::Adapter,
     errors::{AckError, Error},
     ns::Namespace,
     packet::Packet,
@@ -42,6 +42,31 @@ impl<const COUNT: usize> RoomParam for [&'static str; COUNT] {
 
     fn into_room_iter(self) -> Self::IntoIter {
         self.into_iter().map(|s| s.to_string().to_string())
+    }
+}
+
+pub type Room = String;
+
+#[derive(Hash, PartialEq, Eq)]
+pub enum BroadcastFlags {
+    Local,
+    Broadcast,
+    Timeout(Duration),
+}
+pub struct BroadcastOptions {
+    pub flags: HashSet<BroadcastFlags>,
+    pub rooms: Vec<Room>,
+    pub except: Vec<Room>,
+    pub sid: i64,
+}
+impl BroadcastOptions {
+    pub fn new(sid: i64) -> Self {
+        Self {
+            flags: HashSet::new(),
+            rooms: Vec::new(),
+            except: Vec::new(),
+            sid,
+        }
     }
 }
